requirements.txt
Flask==3.1.3
google_api_python_client==2.111.0
gspread==6.2.1
protobuf==6.33.5
python-dotenv==1.2.1
python-telegram-bot==22.6
